Some refactoring done
