+++ app/src/main/java/com/mraulio/gbcameramanager/gbxcart/GBxCartConstants.java
@@ -3,8 +3,7 @@ package com.mraulio.gbcameramanager.gbxcart;
 import java.util.HashMap;
 import java.util.Map;
 
-public class GBxCart {
-
+public class GBxCartConstants {
 
     public static HashMap<String, Integer> DEVICE_CMD = new HashMap<String, Integer>() {{
         put("NULL", 0x30);
