+++ app/src/main/java/com/mraulio/gbcameramanager/ui/gallery/CustomGridViewAdapterImage.java
@@ -1,0 +1,102 @@
+package com.mraulio.gbcameramanager.ui.gallery;
+
+import static android.view.View.GONE;
+
+import android.app.Activity;
+import android.content.Context;
+import android.graphics.Bitmap;
+import android.graphics.Color;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.ImageView;
+import android.widget.TextView;
+
+import com.mraulio.gbcameramanager.R;
+import com.mraulio.gbcameramanager.model.GbcImage;
+import com.mraulio.gbcameramanager.utils.Utils;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CustomGridViewAdapterImage extends ArrayAdapter<GbcImage> {
+    Context context;
+    int layoutResourceId;
+    List<GbcImage> data = new ArrayList<GbcImage>();
+    private List<Bitmap> images;
+    private boolean checkDuplicate;
+    private boolean showName, multiSelect;
+    private List<Integer> selectedImages;
+
+    public CustomGridViewAdapterImage(Context context, int layoutResourceId,
+                                      List<GbcImage> data, List<Bitmap> images, boolean checkDuplicate,
+                                      boolean showName, boolean multiSelect, List<Integer> selectedImages) {
+        super(context, layoutResourceId, data);
+        this.layoutResourceId = layoutResourceId;
+        this.context = context;
+        this.images = images;
+        this.data = data;
+        this.checkDuplicate = checkDuplicate;
+        this.showName = showName;
+        this.multiSelect = multiSelect;
+        this.selectedImages = selectedImages;
+    }
+
+    @Override
+    public View getView(int position, View convertView, ViewGroup parent) {
+        View row = convertView;
+        RecordHolder holder = null;
+        if (row == null) {
+            LayoutInflater inflater = ((Activity) context).getLayoutInflater();
+            row = inflater.inflate(layoutResourceId, parent, false);
+            holder = new RecordHolder();
+            holder.txtTitle = (TextView) row.findViewById(R.id.tvName);
+            holder.imageItem = (ImageView) row.findViewById(R.id.imageView);
+
+            row.setTag(holder);
+        } else {
+            holder = (RecordHolder) row.getTag();
+        }
+        Bitmap image = images.get(position);
+        String name = data.get(position).getName();
+        String hash = data.get(position).getHashCode();
+        List<String> hashToCheck = new ArrayList<>();
+        if (multiSelect && selectedImages != null && !selectedImages.isEmpty()) {
+            for (int i : selectedImages) {
+                hashToCheck.add(GalleryFragment.filteredGbcImages.get(i).getHashCode());
+            }
+            row.setBackgroundColor(hashToCheck.contains(hash) ? context.getColor(R.color.teal_700) : Color.WHITE);
+        }
+
+        Boolean fav = data.get(position).getTags().contains("__filter:favourite__");
+        holder.imageItem.setBackgroundColor(fav ? context.getColor(R.color.favorite) : Color.WHITE);
+        Boolean dup = false;
+        if (checkDuplicate) {
+            for (GbcImage gbcImage : Utils.gbcImagesList) {
+                //Compare the hash value with the value of the new image hash
+                if (gbcImage.getHashCode().equals(hash)) {
+                    //If hash is equals means the image already exists
+                    dup = true;
+                }
+            }
+        }
+        if (showName) {
+            holder.txtTitle.setTextColor(dup ? context.getResources().getColor(R.color.duplicated) : Color.BLACK);
+            if (name.equals("")) {
+                holder.txtTitle.setText("*No title*");
+            } else
+                holder.txtTitle.setText(name);
+        } else {
+            holder.txtTitle.setVisibility(GONE);
+        }
+
+        holder.imageItem.setImageBitmap(Bitmap.createScaledBitmap(image, image.getWidth(), image.getHeight(), false));
+        return row;
+    }
+
+    private class RecordHolder {
+        TextView txtTitle;
+        ImageView imageItem;
+    }
+}
