+++ app/src/main/res/values-de/strings.xml
@@ -140,7 +140,11 @@
     <string name="no_selected">Keine Bilder ausgewählt</string>
     <string name="json_backup_saved">JSON Backup gespeichert!</string>
     <string name="page_selector_dialog">Seite wählen</string>
-    <string name="hdr_exception">Alle Bilder müssen die selbe Größe haben.</string>
+    <string name="hdr_exception">Alle Bilder müssen dieselbe Größe haben.</string>
     <string name="save_location_hint">PNGs, HDRs und Animationen werden nach \n\u0020 Pictures/GBCamera Manager gespeichert\nAndere Dateien werden nach Download/GBCamera Manager gespeichert.</string>
-    
+
+    <string name="done_dumping_ram">RAM WURDE AUSGELESEN! Schau in den Ordner Download/GBCamera Manager/Save dumps</string>
+    <string name="done_dumping_photo">ROM WURDE AUSGELESEN! Schau in den Ordner Download/GBCamera Manager/PHOTO Rom Dumps</string>
+    <string name="filtered_images">Gefilterte Bilder:\u0020</string>
+    <string name="selected_images">Ausgewählt:\u0020</string>
 </resources>
